Move Nitro converter code
